refactor(app): confine daywatch/nightwatch mode indication to the status bar
PR #391 washed the entire window with `tintColor(for:)` at 5% opacity via a
ZStack behind ContentView. Every pane that paints no background of its own
inherited it — the transcript most visibly, sampling (246, 247, 250) where it
should be pure white. That overstepped: mode is a status, not a theme.

- Delete `NightwatchModeTintModifier` and the `.nightwatchModeTint(_:experimentalEnabled:)`
  extension, plus its ContentView call site. `NightwatchModeTheme.tintColor(for:)`
  survives — the chip below is its only consumer now.
- Rework `NightwatchDeskStatusBanner` (a full-width band above the
  NavigationSplitView) into `NightwatchDeskStatusChip`, rendered on the leading
  side of `StatusBarView`'s HStack. Same glyph + text, same tinted
  background/border, same help text and click-to-focus-the-desk behaviour, now
  sized to sit inside the bar (.caption, 2/6pt padding, r=4).
- The chip owns its own `nightwatchExperimentalKey` @AppStorage gate rather
  than relying on the caller — still fail-closed to hidden, and it self-hides
  when mode is .off.
- Drop the chip's `@State deskWorktree` + `.onReceive(appState.objectWillChange)`
  pair, which rescanned `allWorktrees` on every AppState change. The tap action
  resolves the desk on demand instead.
- Copy shortened to "Daywatch desk" / "Nightwatch desk" for the bar; the full
  context stays in the tooltip. Tests updated to match.
- Drop the three `resolvedTint` tests (deleted type); the `tintColor(for:)`
  tests stay.

The menu-bar Nightwatch CommandMenu glyph and the sidebar NightwatchModeToggle
are untouched — the former is menu-bar chrome, not the window, and the latter is
the only in-window control for changing mode.
